Sensor driver added

+++ drivers/iio/adc/at91_adc.c
@@ -30,6 +30,8 @@
 #include <linux/iio/trigger.h>
 #include <linux/iio/trigger_consumer.h>
 #include <linux/iio/triggered_buffer.h>
+//Charm
+#include <linux/Charm/rpc.h>
 
 /* Registers */
 #define AT91_ADC_CR		0x00		/* Control Register */
@@ -182,7 +184,7 @@ struct at91_adc_caps {
 	u8	ts_pen_detect_sensitivity;
 
 	/* startup time calculate function */
-	u32 (*calc_startup_ticks)(u32 startup_time, u32 adc_clk_khz);
+	u32 (*calc_startup_ticks)(u8 startup_time, u32 adc_clk_khz);
 
 	u8	num_channels;
 	struct at91_adc_reg_desc registers;
@@ -201,7 +203,7 @@ struct at91_adc_state {
 	u8			num_channels;
 	void __iomem		*reg_base;
 	struct at91_adc_reg_desc *registers;
-	u32			startup_time;
+	u8			startup_time;
 	u8			sample_hold_time;
 	bool			sleep_mode;
 	struct iio_trigger	**trig;
@@ -780,7 +782,7 @@ ret:
 	return ret;
 }
 
-static u32 calc_startup_ticks_9260(u32 startup_time, u32 adc_clk_khz)
+static u32 calc_startup_ticks_9260(u8 startup_time, u32 adc_clk_khz)
 {
 	/*
 	 * Number of ticks needed to cover the startup time of the ADC
@@ -791,7 +793,7 @@ static u32 calc_startup_ticks_9260(u32 startup_time, u32 adc_clk_khz)
 	return round_up((startup_time * adc_clk_khz / 1000) - 1, 8) / 8;
 }
 
-static u32 calc_startup_ticks_9x5(u32 startup_time, u32 adc_clk_khz)
+static u32 calc_startup_ticks_9x5(u8 startup_time, u32 adc_clk_khz)
 {
 	/*
 	 * For sama5d3x and at91sam9x5, the formula changes to:
